+++ changelog.md
@@ -53,6 +53,10 @@
 * Add Ranked AMQ Alert
 
 
+## (15/03/2024)
+* Add Share button
+* Add go Back button
+
 
 <hr>